# server/openapi.yaml
openapi: 3.0.3
info:
  title: Geopolitical Dashboard API
  version: 1.1.0
  description: |
    Core backend API for datasets, portfolio analytics support data, reports, health checks,
    and news ingestion.
servers:
  - url: http://localhost:5001
paths:
...
  /api/admin/observability:
    get:
      summary: Request-level observability aggregates (admin)
      responses:
        '200':
          description: Request volume, error rate, latency and route stats
        '403':
          description: Forbidden
  /api/admin/alerts:
    get:
      summary: Active server alerts derived from observability thresholds (admin)
      responses:
        '200':
          description: Active alert list and configured thresholds
        '403':
          description: Forbidden
        '429':
          description: Admin abuse protection triggered
  /api/admin/incidents:
    get:
      summary: Recent incidents with tracing context and severity breakdown (admin)
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 500
      responses:
        '200':
          description: Incident timeline and summary
        '403':
          description: Forbidden
        '429':
          description: Admin abuse protection triggered
  /api/admin/audit-trail:
    get:
      summary: Recent admin/release audit events (admin)
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 500
      responses:
        '200':
          description: Recent audit trail entries
        '403':
          description: Forbidden
        '429':
          description: Admin abuse protection triggered
...
  /api/news:
    get:
      summary: Get filtered world news feed
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: News payload
        '429':
          description: Expensive endpoint abuse protection triggered
  /api/telemetry/frontend-crash:
    post:
      summary: Ingest frontend crash telemetry event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '202':
          description: Telemetry event accepted
...
  /api/reports/generate:
    post:
      summary: Generate report metadata
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Generation accepted
        '429':
          description: Expensive endpoint abuse protection triggered
  /api/reports/email:
    post:
      summary: Queue report email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Email queued
        '429':
          description: Expensive endpoint abuse protection triggered
...
  /api/reports/schedule:
    post:
      summary: Schedule recurring report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Schedule created
        '429':
          description: Expensive endpoint abuse protection triggered
...